src/requirements.txt
openai==0.27.6
python-dotenv==1.0.1
tqdm==4.66.5
